Move Redis store logic to config and create sessionStore middleware
