+++ src/server.js
@@ -5,10 +5,6 @@ const compression = require('compression')
 const config = require('../config')
 const express = require('express')
 const flash = require('connect-flash')
-const redis = require('redis')
-const redisCrypto = require('connect-redis-crypto')
-const session = require('express-session')
-const url = require('url')
 const csrf = require('csurf')
 const slashify = require('slashify')
 const churchill = require('churchill')
@@ -24,11 +20,11 @@ const user = require('./middleware/user')
 const auth = require('./middleware/auth')
 const csrfToken = require('./middleware/csrf-token')
 const errors = require('./middleware/errors')
+const sessionStore = require('./middleware/session-store')
 const logger = require('../config/logger')
 
 const router = require('../config/routes')
 
-const isDev = config.isDev
 const app = express()
 
 app.disable('x-powered-by')
@@ -37,59 +33,14 @@ if (!config.ci) {
   app.use(churchill(logger))
 }
 
-if (!isDev) {
+if (!config.isDev) {
   app.enable('trust proxy')
   app.use(enforce.HTTPS({
     trustProtoHeader: true,
   }))
 }
 
-const RedisStore = redisCrypto(session)
-
-let client
-
-if (config.redis.url) {
-  const redisURL = url.parse(config.redis.url)
-  client = redis.createClient(redisURL.port, redisURL.hostname, { no_ready_check: true })
-  client.auth(redisURL.auth.split(':')[1])
-} else {
-  client = redis.createClient(config.redis.port, config.redis.host)
-}
-
-client.on('error', (e) => {
-  logger.log('error', 'Error connecting to redis')
-  logger.log('error', e)
-  throw e
-})
-
-client.on('connect', () => {
-  logger.log('info', 'connected to redis')
-})
-
-client.on('ready', () => {
-  logger.log('info', 'connection to redis is ready to use')
-})
-
-const redisStore = new RedisStore({
-  client,
-  // config ttl defined in milliseconds
-  ttl: config.session.ttl / 1000,
-  secret: config.session.secret,
-})
-
-app.use(session({
-  store: redisStore,
-  proxy: !isDev,
-  cookie: {
-    secure: !isDev,
-    maxAge: config.session.ttl,
-  },
-  rolling: true,
-  key: 'datahub.sid',
-  secret: config.session.secret,
-  resave: true,
-  saveUninitialized: true,
-}))
+app.use(sessionStore)
 
 app.use(bodyParser.urlencoded({ extended: true, limit: '1mb' }))
